feat: make bitfield ops symmetric
And put them in a new file. This is verbose, but should be easy to
understand and reason about.
